simuladores/Simulador-S7/requirements.txt
python-snap7==2.0.2
